Add styling to draggable cards
